backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.2
gunicorn==23.0.0
pydantic[email]==2.11.1
pydantic-settings==2.8.1
python-multipart==0.0.20

# ── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2

# ── Redis ──────────────────────────────────────────────────────
redis[hiredis]==5.2.1

# ── Auth / crypto ──────────────────────────────────────────────
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4

# ── HTTP client ────────────────────────────────────────────────
httpx==0.28.1

# ── Logging / monitoring ──────────────────────────────────────
structlog==25.1.0
sentry-sdk==2.22.0
opentelemetry-sdk==1.31.1
prometheus-client==0.22.0

# ── AI / agents ────────────────────────────────────────────────
langchain==0.3.23
langgraph==0.3.30
anthropic==0.49.0
openai==1.68.2

# ── Supabase ───────────────────────────────────────────────────
supabase==2.13.0

# ── Background jobs ────────────────────────────────────────────
inngest==0.4.21

# ── Vector DB ──────────────────────────────────────────────────
pinecone==6.0.2

# ── Testing ────────────────────────────────────────────────────
pytest==8.3.5
pytest-asyncio==0.25.3
anyio==4.8.0
faker==37.1.0
